Backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-dotenv==1.0.0
google-generativeai>=0.8.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
twilio>=9.0.0
sentence-transformers>=3.3.0
faiss-cpu>=1.9.0
numpy>=1.26.0,<3.0.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.28.0
